requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
streamlit>=1.28.0
pandas>=2.0.0
duckdb>=1.0.0
networkx>=3.0
pyvis>=0.3.0
neo4j>=5.0.0
python-dotenv>=1.0.0
